+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     - devDependencies:
         - @babel/plugin-transform-modules-commonjs: 7.18.6
 - Alerting: attribute "multipleAlert" allows to add single alert or multiple-alert as a list.
+- MapMarker: Added function for rotation in 3D.
 
 ### Changed
 - OAF: